Refactor: 챌린지 모달창 모듈화
