+++ src/pages/MyPage/template/index.tsx
@@ -1,7 +1,6 @@
-import { Modal } from 'components/atoms';
-import { ChallengeModalComment, DeleteModal, MyPageSection } from 'components/molecules';
+import { DeleteModal, MyPageSection } from 'components/molecules';
 import { MyCommentList, MyPageHeader } from 'components/organisms';
-import dayjs from 'dayjs';
+import MyPageCommentModal from 'components/organisms/MyPageCommentModal';
 import React from 'react';
 import { IChallenge, IChallengeData } from 'types/challenge.type';
 import { IConcert } from 'types/concert.type';
@@ -149,37 +148,14 @@ function MyPageTemplate({
         setIsDeleteModalOpen={setIsModalOpened}
         onClickDeleteButton={deleteSelectedCommentList}
       />
-      <Modal isOpen={isChallengeModalOpen} setIsOpen={setIsChallengeModalOpen} isBlur={true}>
-        <S.ChallengeModalWrapper>
-          <S.ModalWrapper>
-            <S.ModalUserInfo>
-              <S.ModalUserImg src={challenge?.user?.img} />
-              <S.ModalUserInfoWrapper>
-                <div>
-                  <div>{challenge?.user?.nickname}</div>
-                  <div>{dayjs().format('MM.DD')}</div>
-                </div>
-                <div>{challenge?.interest?.map((item: string) => `#${item} `)}</div>
-              </S.ModalUserInfoWrapper>
-            </S.ModalUserInfo>
-            <S.TextArea isFolded={isFolded}>
-              <h3>잘한 점</h3>
-              <p>{challenge?.good}</p>
-              <h3>못한 점</h3>
-              <p>{challenge?.bad}</p>
-              <h3>배운 점</h3>
-              <p>{challenge?.learn}</p>
-            </S.TextArea>
-            <S.MoreButton onClick={handleCommentFold}>{isFolded ? '더보기' : '접기'}</S.MoreButton>
-            <S.CommentWrapper>
-              {challenge?.comments.map((comment) => (
-                <ChallengeModalComment key={comment._id} commentData={comment} />
-              ))}
-            </S.CommentWrapper>
-          </S.ModalWrapper>
-          <S.CloseButton onClick={handleCommentFold}>닫기</S.CloseButton>
-        </S.ChallengeModalWrapper>
-      </Modal>
+      <MyPageCommentModal
+        isFolded={isFolded}
+        isChallengeModalOpen={isChallengeModalOpen}
+        challenge={challenge}
+        handleCommentFold={handleCommentFold}
+        setIsChallengeModalOpen={setIsChallengeModalOpen}
+        handleChallengeModalOpen={handleChallengeModalOpen}
+      />
     </S.Wrapper>
   );
 }
